requirements.txt
vkbottle>=4.4,<5
aiohttp>=3.9,<4
python-dotenv>=1.0,<2
